Move (delegate?) course and courseOffering lookups to CourseRepository and move implementations to InMemoryCourseRepository.
