--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/inmemoryrepo/ImMemoryCourseRepository.kt
@@ -1,7 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.inmemoryrepo
-
-import org.springframework.stereotype.Component
-import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.CourseRepository
-
-@Component
-class ImMemoryCourseRepository : CourseRepository
